+++ src/client/core/batchlabs-application.ts
@@ -1,27 +1,33 @@
 import { app, ipcMain, session } from "electron";
 import { AppUpdater, UpdateCheckResult, autoUpdater } from "electron-updater";
 import * as os from "os";
+import * as Url from "url";
 
 import { AuthenticationWindow } from "../authentication";
 import { Constants } from "../client-constants";
 import { logger } from "../logger";
 import { MainWindow } from "../main-window";
+import { PythonRpcServerProcess } from "../python-process";
 import { RecoverWindow } from "../recover-window";
 import { SplashScreen } from "../splash-screen";
 
 const osName = `${os.platform()}-${os.arch()}/${os.release()}`;
 const isDev = Constants.isDev ? "-dev" : "";
-
 const userAgent = `(${osName}) BatchLabs/${Constants.version}${isDev}`;
+
 export class BatchLabsApplication {
     public splashScreen = new SplashScreen(this);
     public authenticationWindow = new AuthenticationWindow(this);
     public recoverWindow = new RecoverWindow(this);
     public mainWindow = new MainWindow(this);
+    public pythonServer = new PythonRpcServerProcess();
 
-    constructor(public autoUpdater: AppUpdater) { }
+    constructor(public autoUpdater: AppUpdater) {
+        logger.info("ARguments", process.argv);
+    }
 
     public init() {
+        this._registerProtocol();
         this.setupProcessEvents();
     }
 
@@ -29,6 +35,7 @@ export class BatchLabsApplication {
      * Start the app by showing the splash screen
      */
     public start() {
+        this.pythonServer.start();
         const requestFilter = { urls: ["https://*", "http://*"] };
         session.defaultSession.webRequest.onBeforeSendHeaders(requestFilter, (details, callback) => {
             if (details.url.indexOf("batch.azure.com") !== -1) {
@@ -43,6 +50,7 @@ export class BatchLabsApplication {
         this.splashScreen.updateMessage("Loading app");
 
         this.mainWindow.create();
+        this._processArguments(process.argv);
     }
 
     public setupProcessEvents() {
@@ -90,10 +98,33 @@ export class BatchLabsApplication {
     }
 
     public quit() {
+        this.pythonServer.stop();
         app.quit();
     }
 
     public checkForUpdates(): Promise<UpdateCheckResult> {
         return autoUpdater.checkForUpdates();
+    }
+
+    private _registerProtocol() {
+        if (Constants.isDev) {
+            return;
+        }
+
+        if (app.setAsDefaultProtocolClient(Constants.customProtocolName)) {
+            logger.info(`Registered ${Constants.customProtocolName}:// as a protocol for batchlabs`);
+        } else {
+            logger.error(`Failed to register ${Constants.customProtocolName}:// as a protocol for batchlabs`);
+        }
+    }
+
+    private _processArguments(argv: string[]) {
+        if (Constants.isDev || argv.length < 2) {
+            return;
+        }
+        const arg = argv[1];
+        if (Url.parse(arg).protocol === Constants.customProtocolName + ":") {
+            this.mainWindow.send(Constants.rendererEvents.batchlabsLink, arg);
+        }
     }
 }