+++ src/client/python-process/python-executable.ts
@@ -41,7 +41,6 @@ export function tryPython(pythonPath: string): Promise<string> {
  * @returns Promise that resolve the first valid path. If no path are valid reject with an error.
  */
 export function tryMultiplePythons(paths: string[]): Promise<string> {
-    const errors = {};
     if (paths.length === 0) {
         return Promise.reject({});
     }