+++ test/app/components/pool/action/add/vm-size-picker.spec.ts
@@ -2,6 +2,7 @@ import { Component, DebugElement, NO_ERRORS_SCHEMA } from "@angular/core";
 import { ComponentFixture, TestBed, fakeAsync, tick } from "@angular/core/testing";
 import { FormsModule } from "@angular/forms";
 import { By } from "@angular/platform-browser";
+import { NoopAnimationsModule } from "@angular/platform-browser/animations";
 import { List } from "immutable";
 import { Observable } from "rxjs";
 
@@ -9,6 +10,7 @@ import { VmSizePickerComponent } from "app/components/pool/action/add";
 import { AccountResource, VmSize } from "app/models";
 import { PoolOsSources } from "app/models/forms";
 import { AccountService, PricingService, VmSizeService } from "app/services";
+import { OSPricing } from "app/services/pricing";
 
 @Component({
     template: `<bl-vm-size-picker [(ngModel)]="vmSize" [osSource]="osSource"></bl-vm-size-picker>`,
@@ -56,11 +58,12 @@ describe("VmSizePickerComponent", () => {
         };
 
         pricingServiceSpy = {
-            getPrices: () => Observable.of([]),
+            getPrice: () => Observable.of(0),
+            getPrices: () => Observable.of(new OSPricing("westus", "linux")),
         };
 
         TestBed.configureTestingModule({
-            imports: [FormsModule],
+            imports: [FormsModule, NoopAnimationsModule],
             declarations: [VmSizePickerComponent, TestComponent],
             providers: [
                 { provide: VmSizeService, useValue: vmSizeServiceSpy },